requirements.txt
python-dotenv
ollama
ollama
psycopg2-binary 
pgvector 
dspy
ftfy 
regex 
tqdm
# git+https://github.com/openai/CLIP.git
git+https://github.com/facebookresearch/ImageBind.git
openai-whisper
opencv-python
transformers 
Pillow
